% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{signature}
\alias{signature}
\title{Signature of a Generalized Ellipsoid}
\usage{
signature(G)
}
\arguments{
\item{G}{A class \code{"gell"} objects}
}
\value{
A vector of length 3, with named components \code{pos}, \code{zero}
           and \code{inf}
}
\description{
Calculates the signature of a generalized ellipsoid, a vector of length 3
giving the number of positive, zero and infinite singular values in the (U,
D) representation
}
\examples{

(zsph <- gell(diag(3)))  # unit sphere in R^3

(zplane <- gell(span = diag(3)[,1:2]))  # a plane

dual(zplane)  # line orthogonal to that plane

(zhplane <- gell(center = c(0,0,2), span = diag(3)[,1:2]))  # a hyperplane

dual(zhplane) # orthogonal line through same center (note that the 'gell'
              # object with a center contains more information than the geometric plane)

zorigin <- gell(span = cbind(c(0,0,0)))
dual( zorigin )

# signatures of these ellipsoids
signature(zsph)
signature(zhplane)
signature(dual(zhplane))

}
\references{
Friendly, M., Monette, G. and Fox, J. (2013). Elliptical
Insights: Understanding Statistical Methods through Elliptical Geometry.
\emph{Statistical Science}, \bold{28}(1), 1-39.
}
\seealso{
\code{\link{isBounded}}
}
\author{
Georges Monette
}
\keyword{dplot}
