% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UD.R
\name{UD}
\alias{UD}
\title{SVD, modified to return U, and D extended with 0's}
\usage{
UD(x)
}
\arguments{
\item{x}{A matrix}
}
\value{
A list with the following components: 
   \item{u}{Right singular vectors} 
   \item{d}{Singular values}
}
\description{
Calculates U and D in the (U, D) representation of a generalized ellipsoid.
This uses the SVD, modified to return U, and D extended with 0's for
singular matrices.
}
\examples{

# None yet

}
\seealso{
\code{\link[base]{svd}}
}
\author{
Georges Monette
}
\keyword{dplot}
