% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isBounded.R
\name{isBounded}
\alias{isBounded}
\alias{isBounded.gell}
\alias{isFat}
\alias{isFat.gell}
\alias{isFlat}
\alias{isUnbounded}
\title{Tests for Classes of Generalized Ellipsoids}
\usage{
isBounded(x, ...)

\method{isBounded}{gell}(x, ...)

isFat(x, ...)

isFlat(x, ...)

isUnbounded(x, ...)
}
\arguments{
\item{x}{A class \code{"gell"} object}

\item{\dots}{Other arguments, not used.}
}
\value{
TRUE or FALSE
}
\description{
These functions provide tests for classes of generalized ellipsoids in the
(U, D) representation, based on the numbers of positive, zero and infinite
singular values in D.
}
\details{
They are included here mainly as computational definitions of the terms
\sQuote{bounded} for an ellipsoid with finite extent, 
\sQuote{fat}, for a bounded ellipsoid with non-empty interior, 
\sQuote{flat}, for degenerate (singular) ellipsoids 
in \eqn{R^p} with empty interior.
}
\examples{

# None yet

}
\references{
Friendly, M., Monette, G. and Fox, J. (2013). Elliptical
Insights: Understanding Statistical Methods through Elliptical Geometry.
\emph{Statistical Science}, \bold{28}(1), 1-39.
}
\seealso{
\code{\link{signature}}
}
\author{
Georges Monette
}
\keyword{logic}
